@@ -3,18 +3,11 @@ using MongoDB.Driver;
 
 namespace ChampionsOfKhazad.Bot.DiscordStats.Mongo;
 
-internal class MongoStreakBreakStore : IStoreStreakBreaks
+internal class MongoStreakBreakStore(IMongoCollection<StreakBreak> streakBreakCollection) : IStoreStreakBreaks
 {
-    private readonly IMongoCollection<StreakBreak> _streakBreakCollection;
-
-    public MongoStreakBreakStore(IMongoCollection<StreakBreak> streakBreakCollection)
-    {
-        _streakBreakCollection = streakBreakCollection;
-    }
-
     public async Task<uint> GetStreakBreakCountByUserAsync(ulong userId, string emoteName, CancellationToken cancellationToken = default)
     {
-        var count = await _streakBreakCollection.CountDocumentsAsync(
+        var count = await streakBreakCollection.CountDocumentsAsync(
             x => x.UserId == userId && x.EmoteName == emoteName,
             cancellationToken: cancellationToken
         );
@@ -22,5 +15,5 @@ internal class MongoStreakBreakStore : IStoreStreakBreaks
         return (uint)count;
     }
 
-    public Task InsertStreakBreakAsync(StreakBreak streakBreak) => _streakBreakCollection.InsertOneAsync(streakBreak);
+    public Task InsertStreakBreakAsync(StreakBreak streakBreak) => streakBreakCollection.InsertOneAsync(streakBreak);
 }
