@@ -3,14 +3,7 @@ using MongoDB.Driver;
 
 namespace ChampionsOfKhazad.Bot.HardcoreStats.Mongo;
 
-internal class MongoCharacterDeathStore : IStoreCharacterDeaths
+internal class MongoCharacterDeathStore(IMongoCollection<CharacterDeath> characterDeathCollection) : IStoreCharacterDeaths
 {
-    private readonly IMongoCollection<CharacterDeath> _characterDeathCollection;
-
-    public MongoCharacterDeathStore(IMongoCollection<CharacterDeath> characterDeathCollection)
-    {
-        _characterDeathCollection = characterDeathCollection;
-    }
-
-    public Task InsertCharacterDeathAsync(CharacterDeath characterDeath) => _characterDeathCollection.InsertOneAsync(characterDeath);
+    public Task InsertCharacterDeathAsync(CharacterDeath characterDeath) => characterDeathCollection.InsertOneAsync(characterDeath);
 }
