@@ -1,13 +1,6 @@
 ﻿namespace ChampionsOfKhazad.Bot.HardcoreStats.CharacterDeaths;
 
-internal class CharacterDeathService : IRecordCharacterDeaths
+internal class CharacterDeathService(IStoreCharacterDeaths deathStore) : IRecordCharacterDeaths
 {
-    private readonly IStoreCharacterDeaths _deathStore;
-
-    public CharacterDeathService(IStoreCharacterDeaths deathStore)
-    {
-        _deathStore = deathStore;
-    }
-
-    public Task RecordCharacterDeathAsync(CharacterDeath characterDeath) => _deathStore.InsertCharacterDeathAsync(characterDeath);
+    public Task RecordCharacterDeathAsync(CharacterDeath characterDeath) => deathStore.InsertCharacterDeathAsync(characterDeath);
 }
