@@ -4,63 +4,52 @@ using Pinecone;
 
 namespace ChampionsOfKhazad.Bot.Lore;
 
-internal class LoreService : IGetLore, IGetRelatedLore, IUpdateLore, ICreateLore
+internal class LoreService(IStoreLore loreStore, EmbeddingsService embeddingsService, IndexService indexService) : IGetLore, IGetRelatedLore, IUpdateLore, ICreateLore
 {
-    private readonly IStoreLore _loreStore;
-    private readonly EmbeddingsService _embeddingsService;
-    private readonly IndexService _indexService;
-
-    public LoreService(IStoreLore loreStore, EmbeddingsService embeddingsService, IndexService indexService)
-    {
-        _loreStore = loreStore;
-        _embeddingsService = embeddingsService;
-        _indexService = indexService;
-    }
-
-    public Task<IReadOnlyList<Lore>> GetLoreAsync(CancellationToken cancellationToken = default) => _loreStore.ReadLoreAsync(cancellationToken);
+    public Task<IReadOnlyList<Lore>> GetLoreAsync(CancellationToken cancellationToken = default) => loreStore.ReadLoreAsync(cancellationToken);
 
     public Task<IReadOnlyList<MemberLore>> GetMemberLoreAsync(CancellationToken cancellationToken = default) =>
-        _loreStore.ReadMemberLoreAsync(cancellationToken);
+        loreStore.ReadMemberLoreAsync(cancellationToken);
 
-    public Task<Lore> GetLoreAsync(string name, CancellationToken cancellationToken = default) => _loreStore.ReadLoreAsync(name, cancellationToken);
+    public Task<Lore> GetLoreAsync(string name, CancellationToken cancellationToken = default) => loreStore.ReadLoreAsync(name, cancellationToken);
 
     public Task<MemberLore> GetMemberLoreAsync(string name, CancellationToken cancellationToken = default) =>
-        _loreStore.ReadMemberLoreAsync(name.Split("-").Last(), cancellationToken);
+        loreStore.ReadMemberLoreAsync(name.Split("-").Last(), cancellationToken);
 
     public async Task UpdateLoreAsync(Lore lore)
     {
-        await _loreStore.UpsertLoreAsync(lore);
+        await loreStore.UpsertLoreAsync(lore);
         await UpsertEmbeddingsAsync(lore);
     }
 
     public async Task UpdateMemberLoreAsync(MemberLore lore)
     {
-        await _loreStore.UpsertMemberLoreAsync(lore);
+        await loreStore.UpsertMemberLoreAsync(lore);
         await UpsertEmbeddingsAsync(lore);
     }
 
     public async Task CreateLoreAsync(params Lore[] lore)
     {
-        await _loreStore.UpsertLoreAsync(lore);
+        await loreStore.UpsertLoreAsync(lore);
         await UpsertEmbeddingsAsync(lore);
     }
 
     public async Task CreateMemberLoreAsync(params MemberLore[] lore)
     {
-        await _loreStore.UpsertMemberLoreAsync(lore);
+        await loreStore.UpsertMemberLoreAsync(lore);
         await UpsertEmbeddingsAsync(lore);
     }
 
     private async Task UpsertEmbeddingsAsync(params IEmbeddable[] embeddables)
     {
         var textEntries = embeddables.Select(x => x.ToTextEntry()).ToArray();
-        var embeddings = await _embeddingsService.CreateEmbeddingsAsync(textEntries);
-        var indexes = await _indexService.ListIndexesAsync();
+        var embeddings = await embeddingsService.CreateEmbeddingsAsync(textEntries);
+        var indexes = await indexService.ListIndexesAsync();
 
         if (!indexes.Contains(Constants.IndexName))
-            await _indexService.CreateIndexAsync(Constants.IndexName);
+            await indexService.CreateIndexAsync(Constants.IndexName);
 
-        var index = await _indexService.GetIndexAsync(Constants.IndexName);
+        var index = await indexService.GetIndexAsync(Constants.IndexName);
         var vectors = embeddings.Select(
             x =>
                 new Vector
@@ -76,10 +65,10 @@ internal class LoreService : IGetLore, IGetRelatedLore, IUpdateLore, ICreateLore
 
     public async Task<IReadOnlyCollection<string>> GetRelatedLoreAsync(string text, uint max = 10)
     {
-        var embeddings = await _embeddingsService.CreateEmbeddingsAsync(new TextEntry("input", text));
+        var embeddings = await embeddingsService.CreateEmbeddingsAsync(new TextEntry("input", text));
         var embedding = embeddings.SingleOrDefault();
 
-        var vectorIndex = await _indexService.GetIndexAsync(Constants.IndexName);
+        var vectorIndex = await indexService.GetIndexAsync(Constants.IndexName);
         var vectors = embedding is not null
             ? await vectorIndex.Query(embedding.Vector, max, includeValues: false, includeMetadata: true)
             : Array.Empty<ScoredVector>();
