@@ -2,22 +2,15 @@
 
 namespace ChampionsOfKhazad.Bot.RaidHelper;
 
-public class RaidHelperClient
+public class RaidHelperClient(HttpClient httpClient)
 {
-    private readonly HttpClient _httpClient;
-
-    public RaidHelperClient(HttpClient httpClient)
-    {
-        _httpClient = httpClient;
-    }
-
     public async Task<EventsResponse> GetEventsAsync(ulong guildId)
     {
-        var response = await _httpClient.GetFromJsonAsync<EventsResponse>($"v3/servers/{guildId}/events");
+        var response = await httpClient.GetFromJsonAsync<EventsResponse>($"v3/servers/{guildId}/events");
 
         return response ?? throw new ApplicationException("Failed to get events");
     }
 
     public Task CreateEventAsync(ulong guildId, ulong channelId, CreateEventRequest request) =>
-        _httpClient.PostAsJsonAsync($"v2/servers/{guildId}/channels/{channelId}/event", request);
+        httpClient.PostAsJsonAsync($"v2/servers/{guildId}/channels/{channelId}/event", request);
 }
