@@ -11,21 +11,12 @@ public record RandomChanceMentionFollowerOptions(
     ulong MentionId
 ) : RandomChanceFollowerOptions(Target, IgnoreBotMentionsInChannelId, Instructions, Chance);
 
-public abstract class RandomChanceMentionFollower : RandomChanceFollower
-{
-    private readonly RandomChanceMentionFollowerOptions _options;
-
-    protected RandomChanceMentionFollower(
-        RandomChanceMentionFollowerOptions options,
+public abstract class RandomChanceMentionFollower(RandomChanceMentionFollowerOptions options,
         Assistant assistant,
         BotContext botContext,
-        ILogger<RandomChanceMentionFollower> logger
-    )
-        : base(options, assistant, botContext, logger)
-    {
-        _options = options;
-    }
-
+        ILogger<RandomChanceMentionFollower> logger)
+    : RandomChanceFollower(options, assistant, botContext, logger)
+{
     protected override bool ShouldTrigger(IUserMessage message) =>
-        message.MentionedUserIds.Contains(_options.MentionId) && base.ShouldTrigger(message);
+        message.MentionedUserIds.Contains(options.MentionId) && base.ShouldTrigger(message);
 }
