@@ -11,21 +11,12 @@ public record RandomChanceTriggerWordFollowerOptions(
     string[] TriggerWords
 ) : RandomChanceFollowerOptions(Target, IgnoreBotMentionsInChannelId, Instructions, Chance);
 
-public abstract class RandomChanceTriggerWordsFollower : RandomChanceFollower
-{
-    private readonly RandomChanceTriggerWordFollowerOptions _options;
-
-    protected RandomChanceTriggerWordsFollower(
-        RandomChanceTriggerWordFollowerOptions options,
+public abstract class RandomChanceTriggerWordsFollower(RandomChanceTriggerWordFollowerOptions options,
         Assistant assistant,
         BotContext botContext,
-        ILogger<RandomChanceTriggerWordsFollower> logger
-    )
-        : base(options, assistant, botContext, logger)
-    {
-        _options = options;
-    }
-
+        ILogger<RandomChanceTriggerWordsFollower> logger)
+    : RandomChanceFollower(options, assistant, botContext, logger)
+{
     protected override bool ShouldTrigger(IUserMessage message) =>
-        _options.TriggerWords.Any(x => message.CleanContent.ToLowerInvariant().Contains(x)) && base.ShouldTrigger(message);
+        options.TriggerWords.Any(x => message.CleanContent.ToLowerInvariant().Contains(x)) && base.ShouldTrigger(message);
 }
