@@ -8,24 +8,17 @@ public record StonerBroFollowerOptions : BaseFollowerOptions
     public const string Key = "StonerBro";
 }
 
-public class StonerBroFollower : RandomChanceFollower
-{
-    public StonerBroFollower(
-        IOptions<AllFollowersOptions> allFollowersOptions,
+public class StonerBroFollower(IOptions<AllFollowersOptions> allFollowersOptions,
         IOptions<StonerBroFollowerOptions> options,
         Assistant assistant,
         BotContext botContext,
-        ILogger<StonerBroFollower> logger
-    )
-        : base(
-            new RandomChanceFollowerOptions(
-                options.Value.ToFollowerTarget(),
-                allFollowersOptions.Value.IgnoreBotMentionsInChannelId,
-                $"You are a stoner bro. You and your friend {options.Value.UserName} are high. You will agree with {options.Value.UserName} and share your shitty philosophical ideas. You will try to encourage {options.Value.UserName} to smoke more weed.",
-                options.Value.Chance
-            ),
-            assistant,
-            botContext,
-            logger
-        ) { }
-}
+        ILogger<StonerBroFollower> logger)
+    : RandomChanceFollower(new RandomChanceFollowerOptions(
+        options.Value.ToFollowerTarget(),
+        allFollowersOptions.Value.IgnoreBotMentionsInChannelId,
+        $"You are a stoner bro. You and your friend {options.Value.UserName} are high. You will agree with {options.Value.UserName} and share your shitty philosophical ideas. You will try to encourage {options.Value.UserName} to smoke more weed.",
+        options.Value.Chance
+    ),
+    assistant,
+    botContext,
+    logger);
