@@ -4,8 +4,4 @@ using Microsoft.Extensions.Options;
 
 namespace ChampionsOfKhazad.Bot;
 
-public class ClownReactor : RandomChanceReactor
-{
-    public ClownReactor(IOptions<ClownReactorOptions> options, ILogger<ClownReactor> logger)
-        : base(new RandomChanceReactorOptions(options.Value.UserId, new[] { new Emoji("🤡") }, 1), logger) { }
-}
+public class ClownReactor(IOptions<ClownReactorOptions> options, ILogger<ClownReactor> logger) : RandomChanceReactor(new RandomChanceReactorOptions(options.Value.UserId, new[] { new Emoji("🤡") }, 1), logger);
