@@ -4,8 +4,4 @@ using Microsoft.Extensions.Options;
 
 namespace ChampionsOfKhazad.Bot;
 
-public class HandOfSalvationReactor : RandomChanceReactor
-{
-    public HandOfSalvationReactor(IOptions<HandOfSalvationReactorOptions> options, ILogger<HandOfSalvationReactor> logger)
-        : base(new RandomChanceReactorOptions(options.Value.UserId, new[] { Emote.Parse("<:salv:1176793539787620383>") }, 1), logger) { }
-}
+public class HandOfSalvationReactor(IOptions<HandOfSalvationReactorOptions> options, ILogger<HandOfSalvationReactor> logger) : RandomChanceReactor(new RandomChanceReactorOptions(options.Value.UserId, new[] { Emote.Parse("<:salv:1176793539787620383>") }, 1), logger);
