@@ -6,23 +6,13 @@ namespace ChampionsOfKhazad.Bot;
 public record RandomChanceReactorOptions(ulong UserId, IEnumerable<IEmote> ReactionEmojis, ushort Chance)
     : GuildMessageReactorOptions(UserId, ReactionEmojis);
 
-public abstract class RandomChanceReactor : GuildMessageReactor
+public abstract class RandomChanceReactor(RandomChanceReactorOptions options, ILogger<RandomChanceReactor> logger) : GuildMessageReactor(options)
 {
-    private readonly RandomChanceReactorOptions _options;
-    private readonly ILogger<RandomChanceReactor> _logger;
-
-    protected RandomChanceReactor(RandomChanceReactorOptions options, ILogger<RandomChanceReactor> logger)
-        : base(options)
-    {
-        _options = options;
-        _logger = logger;
-    }
-
     protected override bool ShouldReact(IUserMessage message)
     {
-        var (success, roll) = RandomUtils.Roll(_options.Chance);
+        var (success, roll) = RandomUtils.Roll(options.Chance);
 
-        _logger.LogInformation(
+        logger.LogInformation(
             "{ReactorType} rolling for message from {Author}: {Roll} - {Result}",
             GetType().Name,
             message.Author.GlobalName ?? message.Author.Username,
